1. Mapping destroy: "Failed to route BeginPlay" error fixed 2. DestroyActorRecursively() "list was modified" error fixed 3. Build errors fixed 4. Multicast delegates removed, events used instead (animation and health widget has custom classes to subscribe events, blueprints updated) 5. Health: heal cup, heal over normal health logic 6. Remote player blueprint removed

+++ Source/LetsGo/Items/HealthItem.cpp
@@ -9,7 +9,7 @@ HealthItem::HealthItem(
 {
 }
 
-int HealthItem::GetHealAmount() const
+float HealthItem::GetHealAmount() const
 {
 	return _healAmount;
 }
